🚑 Workflow Hotfixes
* Fixed issue where the `schedule` event would prematurely deploy the `development` branch
* Fixed multiple compiler-only type errors
